Add class detail and upload functionality with file handling
- Implemented ClassDetailPage to display class information, uploaded questions, and evaluated answers.
- Added UploadFilesPage for teachers to upload answer keys and student answers.
- Integrated file upload service to handle file uploads to Supabase Storage.
- Enhanced error handling and loading states for better user experience.
- Included UI components for displaying upload progress and success messages.
